android-tests: refactor utils folder
