+++ android/DFM/src/test/java/com.darakeon.dfm/api/InternetTest.kt
@@ -1,7 +1,7 @@
 package com.darakeon.dfm.api
 
 import android.os.Build
-import com.darakeon.dfm.utils.MockContext
+import com.darakeon.dfm.utils.activity.MockContext
 import com.darakeon.dfm.tests.setAndroidVersion
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
