+++ android/DFM/src/test/java/com.darakeon.dfm/api/ResponseHandlerTest.kt
@@ -10,15 +10,15 @@ import com.darakeon.dfm.api.entities.Environment
 import com.darakeon.dfm.auth.getValue
 import com.darakeon.dfm.auth.setValue
 import com.darakeon.dfm.base.BaseActivity
-import com.darakeon.dfm.utils.ActivityMock
-import com.darakeon.dfm.utils.CallMock
+import com.darakeon.dfm.utils.activity.ActivityMock
+import com.darakeon.dfm.utils.api.CallMock
 import com.darakeon.dfm.utils.TestException
-import com.darakeon.dfm.utils.assertAlertError
+import com.darakeon.dfm.utils.robolectric.assertAlertError
 import com.darakeon.dfm.utils.execute
-import com.darakeon.dfm.utils.getActivityName
-import com.darakeon.dfm.utils.internetError
-import com.darakeon.dfm.utils.internetSlow
-import com.darakeon.dfm.utils.noBody
+import com.darakeon.dfm.utils.activity.getActivityName
+import com.darakeon.dfm.utils.api.internetError
+import com.darakeon.dfm.utils.api.internetSlow
+import com.darakeon.dfm.utils.api.noBody
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.CoreMatchers.startsWith
 import org.junit.Assert.assertNull
